requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
beautifulsoup4==4.10.0
black==22.1.0
botocore==1.23.46
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.10
click==8.0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
dash==1.21.0
dash-bootstrap-components==0.13.1
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-renderer==1.9.1
dash-table==4.12.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.1
django-allauth==0.49.0
django-appconf==1.0.5
django-bootstrap4==22.1
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-htmx==1.8.0
django-imagekit==4.1.0
django-import-export==2.7.1
django-js-asset==2.0.0
django-on-heroku==1.1.2
django-pandas==0.6.6
django-plotly-dash==1.4.2
django-rest-framework==0.1.0
django-storages==1.12.3
django-templated-mail==1.1.1
django-tinymce==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dpd-components==0.1.0
dpd-static-support==0.0.5
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Compress==1.10.1
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
import-export==0.2.67.dev6
incremental==21.3.0
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupPy==1.14
MarkupSafe==2.0.1
msgpack==1.0.3
mypy-extensions==0.4.3
numpy==1.22.1
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.5
pathspec==0.9.0
patsy==0.5.2
pilkit==2.0
Pillow==9.0.0
platformdirs==2.5.1
plotly==5.6.0
plotly-express==0.4.1
prettytable==3.2.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyecharts==1.9.1
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
s3transfer==0.5.0
scipy==1.8.0
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
statsmodels==0.13.1
tablib==3.1.0
tenacity==8.0.1
tomli==2.0.1
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
uWSGI>=2.0.18,<2.1
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
